+++ Packages/nudge/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.3.0] - 2020-09-24
 ### Added
  - Comments can now reference a list of additional UnityEngine.Objects
  - Shortcuts ctrl shift c and ctrl alt shift c to create Scene Comment and Asset Comment, respectively.